Finished some more updates to the CRUD class.
I started setting up some additional functions to be done w/ the CRUD class to improve its flexibleness and to work on making sure its going to work with all the tables and forms I want to use it for, and get any problems that might require restructuring or major changes to the crud class done because its easier to adjust and change things around a lot if needed when something isn't live yet.
I Extracted part of the Join handling/parsing logic and moved it off into its own function.   This is because the joins and other things can appear in a deep nested format if there are a lot of things like ANDs.   I replaced the main join handling block with a call to this new function.
I Added support for testing if there are bunch of sub expressions in the join members and if so call the new join handling function for each of them, otherwise process it within that join handler call.  With this I can now parse some additional queries and join types.
Started work on various service list alternatives, right now working on a view_vps_list done w/ the CRUD class, which will force me to add support for things like the active/all buttons we use for those, and to finish pagination.


git-svn-id: https://creation.interserver.net/svn/crud/trunk@109 9440e491-775b-45db-8ed1-aeac5ee8fee9