Updated the CRUD JS fixing a rare issue (not sure what causes it yet) that was occasionally getting 'undefined' for the action part of the one form <form action="/yo/url.html?hi=there"> (getting undefined instead of: /yo/url.html?hi=there)
It now checks if it returned undefined for that, and if so now uses the document.location object in js to get the current URL parts and remaps it to the proper/desired ajax page. This fixs the rare problem that comes up but would still like to figure out what initially caused it so i can fix it.